Align package name with dependency reference name
Bender has the issue, that most of the time it ignores the name state it the
bender.yml and uses the name used in dependent packages (e.g. riscv core
package). All other repos refer to this package as "fpnew" although it is
actually called "FPnew". This can cause issues since bender does not
realize that there are clashes between identical packages (according to their
package name) if the parent packages refer to them under different names. This
can cause a whole nightmare of synthesis issues if incompatible versions are
infoluntarely combined that way.